+++ Storage.Azure/AzureStorageService.cs
@@ -15,19 +15,34 @@
         private readonly ILogger<AzureStorageService> _logger;
 
         public AzureStorageService(
-            CloudBlobClient cloudBlobClient,
-            ILogger<AzureStorageService> logger)
+            CloudBlobClient? cloudBlobClient,
+            ILogger<AzureStorageService>? logger)
         {
-            _cloudBlobClient = cloudBlobClient;
-            _logger = logger;
+            _cloudBlobClient = cloudBlobClient ?? throw new ArgumentNullException(nameof(cloudBlobClient));
+            _logger = logger ?? throw new ArgumentNullException(nameof(logger));
         }
 
         public async Task<Uri> Upload(
-            Stream stream,
-            string fileName,
-            string containerName,
+            Stream? stream,
+            string? fileName,
+            string? containerName,
             CancellationToken cancellationToken = default)
         {
+            if (stream == default)
+            {
+                throw new ArgumentNullException(nameof(stream));
+            }
+
+            if (string.IsNullOrEmpty(fileName))
+            {
+                throw new ArgumentNullException(nameof(fileName));
+            }
+
+            if (string.IsNullOrEmpty(nameof(containerName)))
+            {
+                throw new ArgumentNullException(nameof(containerName));
+            }
+
             var container = _cloudBlobClient.GetContainerReference(containerName);
             var blob = container.GetBlockBlobReference(fileName);
             await blob.UploadFromStreamAsync(stream, cancellationToken).ConfigureAwait(false);
@@ -35,19 +50,34 @@
         }
 
         public async Task Delete(
-            string fileName,
-            string containerName,
+            string? fileName,
+            string? containerName,
             CancellationToken cancellationToken = default)
         {
+            if (string.IsNullOrEmpty(fileName))
+            {
+                throw new ArgumentNullException(nameof(fileName));
+            }
+
+            if (string.IsNullOrEmpty(nameof(containerName)))
+            {
+                throw new ArgumentNullException(nameof(containerName));
+            }
+
             var container = _cloudBlobClient.GetContainerReference(containerName);
             var blob = await container.GetBlobReferenceFromServerAsync(fileName, cancellationToken).ConfigureAwait(false);
             await blob.DeleteIfExistsAsync(cancellationToken).ConfigureAwait(false);
         }
 
         public async Task DeleteAll(
-            string containerName,
+            string? containerName,
             CancellationToken cancellationToken = default)
         {
+            if (string.IsNullOrEmpty(nameof(containerName)))
+            {
+                throw new ArgumentNullException(nameof(containerName));
+            }
+
             var container = _cloudBlobClient.GetContainerReference(containerName);
             foreach (var blob in container.ListBlobs(null, true).OfType<CloudBlob>())
             {
@@ -56,16 +86,31 @@
         }
 
         public Uri GetUrl(
-            string fileName,
-            string containerName,
-            DateTime expiration)
+            string? fileName,
+            string? containerName,
+            DateTime? expiration)
         {
+            if (string.IsNullOrEmpty(fileName))
+            {
+                throw new ArgumentNullException(nameof(fileName));
+            }
+
+            if (string.IsNullOrEmpty(nameof(containerName)))
+            {
+                throw new ArgumentNullException(nameof(containerName));
+            }
+
+            if (!expiration.HasValue)
+            {
+                throw new ArgumentNullException(nameof(expiration));
+            }
+
             var container = _cloudBlobClient.GetContainerReference(containerName);
             var blob = container.GetBlockBlobReference(fileName);
             var policy = new SharedAccessBlobPolicy
             {
                 SharedAccessStartTime = DateTimeOffset.UtcNow.AddMinutes(-5),
-                SharedAccessExpiryTime = expiration,
+                SharedAccessExpiryTime = expiration.Value,
                 Permissions = SharedAccessBlobPermissions.Read
             };
             var sharedAccessSignature = blob.GetSharedAccessSignature(policy);
